Trust?

+++ EspelhaML/Domain/MlUserAuthInfo.cs
@@ -4,10 +4,10 @@
     {
         public string AccessToken { get; set; }
         public DateTime ExpiresOn { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public string RefreshToken { get; set; }
 
-        public MlUserAuthInfo(string accessToken, DateTime expiresOn, int userId, string refreshToken)
+        public MlUserAuthInfo(string accessToken, DateTime expiresOn, long userId, string refreshToken)
         {
             AccessToken = accessToken;
             ExpiresOn = expiresOn;
